+++ src/logic/City.ts
@@ -1,7 +1,7 @@
 import { cityConfig } from "./City.config";
 import { CityName } from "./CityName";
 import { ICity } from "./ICity";
-import { Inventory } from "./inventory";
+import { Inventory } from "./Inventory";
 import { IWareForCity } from "./IWareForCity";
 import { WareType } from "./wareType";
